Move layer-specific tests to their own folder.
