+++ run_tests.py
@@ -9,18 +9,18 @@ from tudor import generate_app
 import tests
 from tests.conversions_tests import *
 from tests.convert_task_to_tag_tests import *
-from tests.ll_db_loader_tests import *
+from tests.logic_layer.db_loader_tests import *
 from tests.user_tests import *
-from tests.ll_is_user_authorized_or_admin_tests import *
-from tests.ll_sort_by_hierarchy_tests import *
-from tests.ll_get_index_data_tests import *
-from tests.ll_get_deadlines_data_tests import *
-from tests.ll_create_new_task_tests import *
-from tests.ll_task_set_done_tests import *
-from tests.ll_task_set_deleted_tests import *
-from tests.ll_task_unset_done_tests import *
-from tests.ll_task_unset_deleted_tests import *
-from tests.ll_do_reset_order_nums_tests import *
+from tests.logic_layer.is_user_authorized_or_admin_tests import *
+from tests.logic_layer.sort_by_hierarchy_tests import *
+from tests.logic_layer.get_index_data_tests import *
+from tests.logic_layer.get_deadlines_data_tests import *
+from tests.logic_layer.create_new_task_tests import *
+from tests.logic_layer.task_set_done_tests import *
+from tests.logic_layer.task_set_deleted_tests import *
+from tests.logic_layer.task_unset_done_tests import *
+from tests.logic_layer.task_unset_deleted_tests import *
+from tests.logic_layer.do_reset_order_nums_tests import *
 from tests.tag_tests import *
 from tests.task_tests import *
 from tests.task_id_tests import *
@@ -28,17 +28,17 @@ from tests.task_css_tests import *
 from tests.task_tags_tests import *
 from tests.task_duration_text_tests import *
 from tests.task_cost_text_tests import *
-from tests.ll_search_tests import *
+from tests.logic_layer.search_tests import *
 from tests.task_dependencies_tests import *
-from tests.ll_task_dependencies_tests import *
-from tests.ll_task_tags_tests import *
-from tests.ll_get_lowest_highest_order_num import *
+from tests.logic_layer.task_dependencies_tests import *
+from tests.logic_layer.task_tags_tests import *
+from tests.logic_layer.get_lowest_highest_order_num import *
 from tests.task_prioritize_tests import *
-from tests.ll_task_prioritize_tests import *
-from tests.vl_route_tests import *
+from tests.logic_layer.task_prioritize_tests import *
+from tests.view_layer.route_tests import *
 from tests.persistence_layer_tests import *
-from tests.ll_do_import_data_tests import *
-from tests.ll_set_task_tests import *
+from tests.logic_layer.do_import_data_tests import *
+from tests.logic_layer.set_task_tests import *
 from tests.task_interlinking_tests import *
 from tests.interlinking_tests import *
 from tests.task_from_dict_tests import *
