+++ tests/logic_layer/task_set_deleted_tests.py
